+++ crates/goose/src/acp/transport/mod.rs
@@ -1,4 +1,6 @@
 pub mod auth;
+#[cfg(any(feature = "rustls-tls", feature = "native-tls"))]
+pub mod tls;
 
 use std::sync::Arc;
